+++ R/eem_correction.R
@@ -1,11 +1,15 @@
 #' Blank correction
 #'
-#' @description This function is used to remove blank from eems which can help to reduce the
-#' effect of scatter bands.
+#' @description This function is used to remove blank from eems which can help
+#'   to reduce the effect of scatter bands.
 #'
 #' @template template_eem
 #' @template template_blank
 #' @template template_details_automatic_blank
+#'
+#' @details Note that blank correction should be performed before Raman
+#'   normalization (\code{eem_raman_normalisation()}). An error will occur
+#'   if trying to perform blank correction after Raman normalization.
 #'
 #' @references Murphy, K. R., Stedmon, C. a., Graeber, D., & Bro, R. (2013).
 #'   Fluorescence spectroscopy and multi-way techniques. PARAFAC. Analytical
@@ -69,7 +73,16 @@ eem_remove_blank <- function(eem, blank = NA) {
   stopifnot(.is_eemlist(eem) | .is_eem(eem),
             .is_eemlist(blank) | is.na(blank))
 
-  if(is.na(blank)){
+  is_raman_normalized <- lapply(eem,
+                               function(x){attributes(x)$is_raman_normalized})
+  is_raman_normalized <- unlist(is_raman_normalized)
+
+  if (any(is_raman_normalized)) {
+    stop("Samples have been Raman normalized. Please perform blank removal
+         before Raman normalization.", call. = FALSE)
+  }
+
+  if (is.na(blank)) {
 
     t <- list.group(eem, ~location)
     t <- lapply(t, function(x){class(x) <- "eemlist"; return(x)})
@@ -90,25 +103,19 @@ eem_remove_blank <- function(eem, blank = NA) {
             .is_eemlist(blank) | is.na(blank))
 
   ## It is a list of eems, then call lapply
-  if(.is_eemlist(eem)){
-
-    blank_names <- c("nano", "miliq", "milliq", "mq", "blank")
+  if (.is_eemlist(eem)) {
 
     # if blank is NA then try to split the eemlist into blank and eems
-    if(is.na(blank)){
-      blank <- eem_extract(eem, blank_names, remove = FALSE, ignore_case = TRUE,
-                           verbose = FALSE)
-      eem <- eem_extract(eem, blank_names, remove = TRUE, ignore_case = TRUE,
-                         verbose = FALSE)
+    if (is.na(blank)) {
 
-      if(length(blank) != 1 | length(eem) < 1){
+      blank <- eem_extract_blank(eem)
+
+      if (length(blank) != 1 | length(eem) < 1) {
         stop("Cannot find blank for automatic correction.", call. = FALSE)
       }
     }
 
-    res <- lapply(eem,
-                  eem_remove_blank,
-                  blank = blank)
+    res <- lapply(eem, eem_remove_blank, blank = blank)
 
     class(res) <- class(eem)
     return(res)
@@ -117,14 +124,22 @@ eem_remove_blank <- function(eem, blank = NA) {
   #---------------------------------------------------------------------
   # Do the blank subtraction.
   #---------------------------------------------------------------------
+
+  # Do not correct if it was already done
+  if(attributes(eem)$is_blank_corrected) { return(eem) }
+
+  if (is_blank(eem)) { return(eem) } # do not modify blank samples
+
   blank <- unlist(blank, recursive = FALSE)
+
   x <- eem$x - blank$x
 
   ## Construct an eem object.
   res <- eem(file = eem$sample,
              x = x,
              ex = eem$ex,
-             em = eem$em)
+             em = eem$em,
+             location = eem$location)
 
   attributes(res) <- attributes(eem)
   attr(res, "is_blank_corrected") <- TRUE
@@ -255,7 +270,8 @@ eem_remove_scattering <- function(eem, type, order = 1, width = 10){
 #' @details The normalization procedure consists in dividing all fluorescence
 #'   intensities by the area (integral) of the Raman peak. The peak is located
 #'   at excitation of 350 nm. (ex = 370) betwen 371 nm. and 428 nm in emission
-#'   (371 <= em <= 428).
+#'   (371 <= em <= 428). Note that the data is interpolated to make sure that
+#'   fluorescence at em 350 exist.
 #'
 #' @references
 #'
@@ -321,24 +337,18 @@ eem_raman_normalisation <- function(eem, blank = NA) {
   ## It is a list of eems, then call lapply
   if(.is_eemlist(eem)){
 
-    blank_names <- c("nano", "miliq", "milliq", "mq", "blank")
-
     # if blank is NA then try to split the eemlist into blank and eems
     if(is.na(blank)){
 
-      blank <- eem_extract(eem, blank_names, remove = FALSE, ignore_case = TRUE,
-                           verbose = FALSE)
-      eem <- eem_extract(eem, blank_names, remove = TRUE, ignore_case = TRUE,
-                         verbose = FALSE)
+      blank <- eem_extract_blank(eem)
 
       if(length(blank) != 1 | length(eem) < 1){
         stop("Cannot find blank for automatic correction.", call. = FALSE)
+
       }
     }
 
-    res <- lapply(eem,
-                  .eem_raman_normalisation,
-                  blank = blank)
+    res <- lapply(eem, .eem_raman_normalisation, blank = blank)
 
     class(res) <- class(eem)
     return(res)
@@ -347,18 +357,30 @@ eem_raman_normalisation <- function(eem, blank = NA) {
   #---------------------------------------------------------------------
   # Do the normalisation.
   #---------------------------------------------------------------------
+
+  # Do not correct if it was already done
+  if(attributes(eem)$is_raman_normalized) { return(eem) }
+
+  # Do not modify blank samples
+  if(is_blank(eem)) {return(eem)}
+
   blank <- unlist(blank, recursive = FALSE)
-  index_ex <- which(blank$ex == 350)
-  index_em <- which(blank$em >= 371 & blank$em <= 428)
 
-  x <- blank$em[index_em]
-  y <- blank$x[index_em, index_ex]
+  em <- seq(371, 428, by = 2)
+  ex <- rep(350, length(em))
+  fluo <- pracma::interp2(blank$ex, blank$em, blank$x, ex, em)
 
-  if(any(is.na(x)) | any(is.na(y))){
+  # index_ex <- which(blank$ex == 350)
+  # index_em <- which(blank$em >= 371 & blank$em <= 428)
+  #
+  # x <- blank$em[index_em]
+  # y <- blank$x[index_em, index_ex]
+
+  if(any(is.na(em)) | any(is.na(fluo))){
     stop("NA values found in the blank sample. Maybe you removed scattering too soon?", call. = FALSE)
   }
 
-  area <- sum(diff(x) * (y[-length(y)] + y[-1]) / 2)
+  area <- sum(diff(em) * (fluo[-length(fluo)] + fluo[-1]) / 2)
 
   cat("Raman area:", area, "\n")
 
@@ -368,7 +390,8 @@ eem_raman_normalisation <- function(eem, blank = NA) {
   res <- eem(file = eem$sample,
              x = x,
              ex = eem$ex,
-             em = eem$em)
+             em = eem$em,
+             location = eem$location)
 
   attributes(res) <- attributes(eem)
   attr(res, "is_raman_normalized") <- TRUE
@@ -454,7 +477,8 @@ eem_inner_filter_effect <- function(eem, absorbance, pathlength = 1) {
   #---------------------------------------------------------------------
   # Some checks
   #---------------------------------------------------------------------
-  names(absorbance) <- tolower(names(absorbance))
+
+  # names(absorbance) <- tolower(names(absorbance))
 
   if(!any(names(absorbance) == "wavelength")){
 
@@ -492,6 +516,9 @@ eem_inner_filter_effect <- function(eem, absorbance, pathlength = 1) {
   # Create the ife matrix
   #---------------------------------------------------------------------
 
+  # Do not correct if it was already done
+  if(attributes(eem)$is_ife_corrected) { return(eem) }
+
   sf <- stats::splinefun(wl, spectra)
 
   ex <- sf(eem$ex)
@@ -516,10 +543,13 @@ eem_inner_filter_effect <- function(eem, absorbance, pathlength = 1) {
   x <- eem$x / ife_correction_factor
 
   ## Construct an eem object.
-  res <- eem(file = eem$sample,
-             x = x,
-             ex = eem$ex,
-             em = eem$em)
+  res <- eem(
+    file = eem$sample,
+    x = x,
+    ex = eem$ex,
+    em = eem$em,
+    location = eem$location
+  )
 
   attributes(res) <- attributes(eem)
   attr(res, "is_ife_corrected") <- TRUE
@@ -532,3 +562,12 @@ is_between <- function(x, a, b) {
   x >= a & x <= b
 }
 
+# Return TRUE if eem is a blank sample
+
+is_blank <- function(eem) {
+
+  blank_names <- paste("nano", "miliq", "milliq", "mq", "blank", sep = "|")
+
+  grepl(blank_names, eem$sample, ignore.case = TRUE)
+
+}