+++ R/eem_utils.R
@@ -1,37 +1,17 @@
-#' Surface plot of eem
-#'
-#' @param x An object of class \code{eemlist}.
-#' @param ... Extra arguments for \code{image.plot}.
-#' @param show_peaks Boolean indicating if Cobble's peaks should be displayed on
-#'   the surface plot. Default is FALSE.
 #' @importFrom grDevices colorRampPalette
-#' @export
-#' @examples
-#' file <- system.file("extdata/cary/scans_day_1", "sample1.csv", package = "eemR")
-#' eem <- eem_read(file)
-#'
-#' plot(eem)
+#' @importFrom graphics par plot text
+#' @importFrom viridis viridis
+.plot_eem <- function(x, show_peaks, ...){
 
-plot.eem <- function(x, show_peaks, ...){
 
-  jet.colors <- colorRampPalette(c("#00007F",
-                                   "blue",
-                                   "#007FFF",
-                                   "cyan",
-                                   "#7FFF7F",
-                                   "yellow",
-                                   "#FF7F00",
-                                   "red",
-                                   "#7F0000"))
-
-  fields::image.plot(y = x$em,
+fields::image.plot(y = x$em,
              x = x$ex,
              z = t(x$x),
              main = paste(x$sample, "\n", attr(x, "manucafturer"), sep = ""),
              xlab = "Excitation (nm.)",
              ylab = "Emission (nm.)",
              legend.lab = "Fluorescence intensity",
-             col = jet.colors(255),
+             col = viridis::viridis(256),
              ...)
 
   if(show_peaks){
@@ -55,6 +35,7 @@ plot.eem <- function(x, show_peaks, ...){
 #' @param ... Extra arguments for \code{image.plot}.
 #' @param show_peaks Boolean indicating if Cobble's peaks should be displayed on
 #'   the surface plot. Default is FALSE.
+#' @param interactive If \code{TRUE} a Shiny app will start to visualize EEMS.
 #'
 #' @export
 #' @examples
@@ -62,11 +43,16 @@ plot.eem <- function(x, show_peaks, ...){
 #' eem <- eem_read(folder)
 #'
 #' plot(eem, which = 3)
-plot.eemlist <- function(x, which = 1, show_peaks = FALSE, ...) {
+plot.eemlist <- function(x, which = 1,
+                         interactive = FALSE, show_peaks = FALSE, ...) {
 
   stopifnot(which <= length(x))
 
-  plot.eem(x[[which]], show_peaks, ...)
+  if(interactive){
+    .plot_shiny(x)
+  }else{
+    .plot_eem(x[[which]], show_peaks, ...)
+  }
 
 }
 
@@ -74,9 +60,7 @@ plot.eemlist <- function(x, which = 1, show_peaks = FALSE, ...) {
 #'
 #' @param object An object of class \code{eem}.
 #' @param ... Extra arguments.
-#'
-#' @references \url{http://www.sciencedirect.com/science/article/pii/0304420395000623}
-#' @importFrom utils head tail
+#' @template template_summary
 #' @export
 #' @examples
 #' file <- system.file("extdata/cary/scans_day_1/", "sample1.csv", package = "eemR")
@@ -88,39 +72,55 @@ summary.eem <- function(object, ...){
 
   stopifnot(class(object) == "eem")
 
-  cat("eem object:", dim(object$x)[1],
-      "x",  dim(object$x)[2],
-      "(", dim(object$x)[1] * dim(object$x)[2], ")", "\n")
+  df <- data.frame(
+    sample = object$sample,
+    ex_min = min(object$ex),
+    ex_max = max(object$ex),
+    em_min = min(object$em),
+    em_max = max(object$em),
+    is_blank_corrected = attr(object, "is_blank_corrected"),
+    is_scatter_corrected = attr(object, "is_scatter_corrected"),
+    is_ife_corrected = attr(object, "is_ife_corrected"),
+    is_raman_normalized = attr(object, "is_raman_normalized"),
+    manufacturer = attr(object, "manufacturer")
+    )
 
-  cat("ex: (", range(object$ex), "nm.)", head(object$ex, 3), "...", tail(object$ex, 3), "\n")
+  return(df)
 
-  cat("em: (", range(object$em), "nm.)", head(object$em, 3), "...", tail(object$em, 3), "\n")
-
-  cat("is_blank_corrected:", attr(object, "is_blank_corrected"), "\n")
-
-  cat("is_scatter_corrected:", attr(object, "is_scatter_corrected"), "\n")
-
-  cat("is_ife_corrected:", attr(object, "is_ife_corrected"), "\n")
-
-  cat("is_raman_normalized:", attr(object, "is_raman_normalized"), "\n")
-
-  cat("manucafturer:", attr(object, "manucafturer"), "\n")
 }
 
-print.eem <- function(object, ...){
-  summary(object)
+print.eem <- function(x, ...){
+  summary(x)
 }
 
-print.eemlist <- function(object, ...){
-  summary(object)
-}
+#' Display summary of an eemlist object
+#'
+#' @param x An object of class \code{eemlist}.
+#' @param ... Extra arguments.
+#' @template template_summary
+#'
+#' @export
+#' @examples
+#' folder <- system.file("extdata/cary", package = "eemR")
+#' eem <- eem_read(folder, recursive = TRUE)
+#'
+#' print(eem)
+print.eemlist <- function(x, ...){
+  stopifnot(class(x) == "eemlist")
 
+  df <- lapply(x, summary)
+  df <- do.call(rbind, df)
+
+  print(df)
+
+  invisible(df)
+}
 
 #' Display summary of an eemlist object
 #'
 #' @param object An object of class \code{eemlist}.
 #' @param ... Extra arguments.
-#'
+#' @template template_summary
 #' @export
 #' @examples
 #' folder <- system.file("extdata/cary", package = "eemR")
@@ -131,10 +131,11 @@ summary.eemlist <- function(object, ...){
 
   stopifnot(class(object) == "eemlist")
 
-  cat("eemlist object containing:", length(object), "eem\n\n")
+  df <- lapply(object, summary)
+  df <- do.call(rbind, df)
 
-  cat("First eem object:\n\n")
-  summary.eem(object[[1]])
+  return(df)
+
 }
 
 
@@ -146,7 +147,6 @@ summary.eemlist <- function(object, ...){
 #' @param fill_with_na Logical. If TRUE, fluorescence values at specified
 #'   wavelengths will be replaced with NA. If FALSE, these values will be
 #'   removed.
-#'
 #' @export
 #' @examples
 #' # Open the fluorescence eem
@@ -157,6 +157,10 @@ summary.eemlist <- function(object, ...){
 #'
 #' # Cut few excitation wavelengths
 #' eem <- eem_cut(eem, ex = c(220, 225, 230, 230))
+#' plot(eem)
+#'
+#' eem <- eem_read(file)
+#' eem <- eem_cut(eem, em = 350:400, fill_with_na = TRUE)
 #' plot(eem)
 eem_cut <- function(eem, ex, em, fill_with_na = FALSE){
 
@@ -196,7 +200,6 @@ eem_cut <- function(eem, ex, em, fill_with_na = FALSE){
         eem$x <- eem$x[, -index]
       }
     }
-
   }
 
   if(!missing(em)){
@@ -499,3 +502,135 @@ my_unlist <- function(x){
 .is_eem <- function(eem) {
   ifelse(class(eem) == "eem", TRUE, FALSE)
 }
+
+.plot_shiny <- function(eem){
+
+  metrics <- dplyr::left_join(eem_coble_peaks(eem, verbose = FALSE),
+                              eem_biological_index(eem, verbose = FALSE),
+                              by = "sample")
+
+  metrics <- dplyr::left_join(metrics,
+                              eem_fluorescence_index(eem, verbose = FALSE),
+                              by = "sample")
+
+  metrics <- dplyr::left_join(metrics,
+                              eem_humification_index(eem, verbose = FALSE),
+                              by = "sample")
+
+  metrics[,-1] <-round(metrics[,-1], digits = 2)
+
+  # nl <- vector(mode = "list", length = length(eem_names(eem)))
+  # names(nl) <- eem_names(eem)
+  # nl[1:length(nl)] <- 1:length(nl)
+
+  ui <- shiny::fluidPage(
+
+    shiny::titlePanel("EEM interactive visualization"),
+
+    shiny::sidebarLayout(
+      shiny::sidebarPanel
+      (
+        shiny::checkboxInput("scale", label = "Keep z-axis fixed?", FALSE),
+        shiny::hr(),
+        shiny::checkboxInput("by", "Combined 2x2 plots", FALSE),
+        shiny::hr(),
+        shiny::sliderInput("ex_cut", "Select excitation range",
+                           min = min(eem[[1]]$ex),
+                           max = max(eem[[1]]$ex),
+                           value = c(min(eem[[1]]$ex), max(eem[[1]]$ex)),
+                           step = 1),
+        shiny::hr(),
+        shiny::sliderInput("em_cut", "Select emission range",
+                           min = min(eem[[1]]$em),
+                           max = max(eem[[1]]$em),
+                           value = c(min(eem[[1]]$em), max(eem[[1]]$em)),
+                           step = 1)
+      ),
+
+
+      shiny::mainPanel(shiny::plotOutput(outputId = "myeem", width = "550px", height = "550px"))
+    ),
+
+    shiny::br(),
+
+    DT::dataTableOutput("eem_list"),
+
+    shiny::br()
+
+)
+
+  server <- function(input, output) {
+
+    output$myeem <- shiny::renderPlot({
+
+      if(input$scale){
+        zlim <- range(unlist(lapply(eem, function(x) x$x)), na.rm = TRUE)
+      } else {
+        zlim <- range(eem[[input$eem_list_rows_selected]]$x, na.rm = TRUE)
+      }
+
+      if(!is.null(input$eem_list_rows_selected)){
+
+        n <- ifelse(input$by, 2, 1)
+
+        par(mfrow = c(n, n))
+
+        plot(eem,
+             which = input$eem_list_rows_selected,
+             xlim = c(input$ex_cut[1], input$ex_cut[2]),
+             ylim = c(input$em_cut[1], input$em_cut[2]),
+             zlim = zlim)
+
+      }
+    })
+
+    output$eem_list = DT::renderDataTable(
+      metrics,
+      server = FALSE,
+      selection = 'single',
+      # selection = list(mode = 'single', target = "row", selected = c(1)),
+      options = list(
+        autoWidth = TRUE,
+        columnDefs = list(list(width = '10px', targets = "_all"))
+      )
+
+    )
+  }
+
+  shiny::shinyApp(ui, server)
+}
+
+
+#' Extract blank EEM
+#'
+#' @template template_eem
+#' @param average Logical. If TRUE blank EEMs will be averaged
+eem_extract_blank <- function(eem, average = TRUE) {
+
+  blank_names <- c("nano", "miliq", "milliq", "mq", "blank")
+
+  blank <- eem_extract(eem, blank_names,
+                       remove = FALSE,
+                       ignore_case = TRUE,
+                       verbose = FALSE)
+
+  # Average all the blank EEMs
+  if(average) {
+
+    n <- length(blank)
+
+    message("A total of ", n, " blank EEMs will be averaged.")
+
+    X <- Reduce("+", lapply(blank, function(x) x$x))
+    X <- X / n
+
+    blank <- blank[1]
+    blank[[1]]$x <- X
+
+    class(blank) <- "eemlist"
+
+  }
+
+  return(blank)
+}
+