+++ man-roxygen/template_details_automatic_blank.R
@@ -8,24 +8,16 @@
 #'   Note that if \code{blank} is omited, the function will group the
 #'   \code{eemlist} based on file location and will assumes that there is a
 #'   blank sample in each folder. In that context, the blank will be used on
-#'   each sample in the same folder.
+#'   each sample in the same folder. If more than one blank is found they will
+#'   be averaged (a message will be printed if this appends).
 #'
 #'   Consider the following example where there are two folders that could
 #'   represent scans performed on two different days `scans_day_1` and
 #'   `scans_day_2`.
 #'
-#'   \tabular{ll}{
-#'   scans_day_1\tab\cr
-#'   \tab nano.csv\cr
-#'   \tab sample1.csv\cr
-#'   \tab sample2.csv\cr
-#'   \tab sample3.csv\cr
-#'   scans_day_2 \tab\cr
-#'   \tab blank.csv\cr
-#'   \tab s1.csv\cr
-#'   \tab s2.csv\cr
-#'   \tab s3.csv\cr
-#'   }
+#'   \tabular{ll}{ scans_day_1\tab\cr \tab nano.csv\cr \tab sample1.csv\cr \tab
+#'   sample2.csv\cr \tab sample3.csv\cr scans_day_2 \tab\cr \tab blank.csv\cr
+#'   \tab s1.csv\cr \tab s2.csv\cr \tab s3.csv\cr }
 #'
 #'   In each folder there are three samples and one blank files. In that
 #'   context, `eem_remove_blank()` will use the blank `nano.csv` from