+++ tests/testthat/test-blank-correction.R
@@ -7,9 +7,12 @@ test_that("grouped blank correction works", {
 
   eems <- eem_remove_blank(eems)
 
+  mysum <- sum(eem_extract(eems, "sample1")[[1]]$x)
+  mymean <- mean(eem_extract(eems, "sample1")[[1]]$x)
+
   # These values have been calculated by "hand" in Excel.
-  expect_equal(sum(eems[[1]]$x), 14020.78844, tolerance = 0.00001)
-  expect_equal(mean(eems[[1]]$x), 1.6038, tolerance = 0.0001)
+  expect_equal(mysum, 14020.78844, tolerance = 0.00001)
+  expect_equal(mymean, 1.6038, tolerance = 0.0001)
 
 })
 
@@ -24,8 +27,11 @@ test_that("single blank correction works", {
 
   eems <- eem_remove_blank(eems, blank)
 
+  mysum <- sum(eem_extract(eems, "sample1")[[1]]$x)
+  mymean <- mean(eem_extract(eems, "sample1")[[1]]$x)
+
   # These values have been calculated by "hand" in Excel.
-  expect_equal(sum(eems[[1]]$x), 14020.78844, tolerance = 0.00001)
-  expect_equal(mean(eems[[1]]$x), 1.6038, tolerance = 0.0001)
+  expect_equal(mysum, 14020.78844, tolerance = 0.00001)
+  expect_equal(mymean, 1.6038, tolerance = 0.0001)
 
 })